pyproject.toml
[project]
name = "card-inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "google-genai>=1.39.1",
    "ipython>=9.6.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
